add object resource and add some test data
Currently, no real database is used, and the test data represents
a small world with a "cross roads" and a couple of street lamps.

It is enough to create the REST API queries for all objects within
a region, and writing the code on the client side for building
the 3D scene of (a portion of) the world.